requirements.txt
catboost
einops==0.8.1
loguru
matplotlib
mlxtend==0.23.4
numpy
pandas
plotly==6.1.2
prophe
pydantic==2.11.7
pyfunctional==1.5.0
scikit_learn
scipy
seaborn
setuptools==78.1.1
statsmodels
torch==2.7.1
torchmetrics==1.7.3
tqdm
